+++ backend/targetusers/views.py
@@ -1,13 +1,12 @@
-from rest_framework import status
-from rest_framework.response import Response
 from rest_framework.viewsets import ModelViewSet
 
+from core.views import CreateByNotificationTypeMixin
 from notifications.models import EnumNotificationType
 from targetusers.models import TargetUser
 from targetusers.serializers import TargetUserSerializer, SlackTargetUserSerializer
 
 
-class TargetUserViewSet(ModelViewSet):
+class TargetUserViewSet(CreateByNotificationTypeMixin, ModelViewSet):
     queryset = TargetUser.objects.all()
     serializer_class = TargetUserSerializer
 
@@ -15,17 +14,6 @@ class TargetUserViewSet(ModelViewSet):
         if project_id := request.query_params.get('projectId'):
             self.queryset = self.queryset.filter(project_id=int(project_id))
         return super().list(request, *args, **kwargs)
-
-    # pylint: disable=R0801
-    def create(self, request, *args, **kwargs):
-        if not ('notification_type' in request.data and
-                request.data['notification_type'] in EnumNotificationType):
-            return Response(status=status.HTTP_400_BAD_REQUEST)
-        serializer_class = self.get_create_serializer_class()
-        serializer = serializer_class(data=request.data)
-        serializer.is_valid(raise_exception=True)
-        self.perform_create(serializer)
-        return Response(serializer.data, status=status.HTTP_201_CREATED)
 
     # pylint: disable=inconsistent-return-statements
     def get_create_serializer_class(self):
